fix(search): responsive and accent
