+++ src/components/atoms/icons/index.ts
@@ -24,4 +24,4 @@ export { default as LogoDisMoiWithD } from './LogoDisMoiWithD';
 export { default as Bullito } from './Bullito';
 export { default as Reload } from './Reload';
 export { default as Play } from './Play';
-export { default as Search } from './Search';
+export { default as SearchIcon } from './Search';
