+++ src/components/molecules/Search/index.ts
@@ -1,0 +1,3 @@
+import Search from './Search';
+
+export default Search;
